Backend/cliente_app/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
python-decouple==3.8
Pillow==11.0.0
openai==1.57.2
numpy==2.2.0
scikit-learn==1.6.0
stripe==11.2.0
